TubeRadioSim V1.0
Version 1 of my Arduino Tube AM radio simulator. The project connects to the mechanical on/off control and volume control within an old radio. It uses NeoPixels to simulate  the warm up time and color of the old radio's pilot lamps and vacuum tubes and Bluetooth for sound.

It uses the following hardware: 
#   EspressIF ESP32-PICO-Kit D4
#   Adafruit MAX98357A I2S class-D amplifier
#   Five Adafruit RGBW NeoPixels
#   LiPoly power supply / charger (my design)
#   Bluetooth Aruino libraries from: Tierneytim - https://github.com/tierneytim/btAudio
#   Read more in the accompanying article in Design News, by Duane Benson